requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.3.0
Markdown==3.4.3
pycodestyle==2.10.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==2.0.2
zipp==3.15.0
